services_new/license_plate_recognition/requirements.txt
opencv-python==4.12.0.88
pytesseract==0.3.13
pillow==11.3.0
numpy==2.0.2
piexif==1.1.3
redis==6.3.0
ultralytics==8.3.175
